pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transaction-classifier"
version = "1.0.0"
description = "Account code prediction for French financial transactions"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1.0,<3",
    "numpy>=1.26.0,<2",
    "scipy>=1.12.0,<2",
    "scikit-learn>=1.4.0,<2",
    "xgboost>=2.0.0,<3",
    "beautifulsoup4>=4.12.0,<5",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "joblib>=1.3.0,<2",
    "pyyaml>=6.0,<7",
    "matplotlib>=3.8.0,<4",
]

[project.optional-dependencies]
train = ["click>=8.1.0,<9", "psycopg2-binary>=2.9.0,<3"]
serve = ["fastapi>=0.115.0,<1", "uvicorn[standard]>=0.34.0,<1", "click>=8.1.0,<9"]
hpo = ["optuna>=4.0,<5", "optuna-integration>=4.0,<5", "plotly>=6.0,<7", "kaleido>=0.2"]
explain = ["shap>=0.46,<1"]
all = [
    "click>=8.1.0,<9", "psycopg2-binary>=2.9.0,<3",
    "fastapi>=0.115.0,<1", "uvicorn[standard]>=0.34.0,<1",
    "optuna>=4.0,<5", "optuna-integration>=4.0,<5", "plotly>=6.0,<7", "kaleido>=0.2",
    "shap>=0.46,<1",
]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-cov>=5.0.0,<6",
    "pytest-asyncio>=0.24.0,<1",
    "httpx>=0.28.0,<1",
    "ruff>=0.4.0,<1",
    "mypy>=1.8.0,<2",
    "pre-commit>=3.7,<4",
    "pandas-stubs>=3.0",
    "types-pyyaml>=6.0",
    "types-psycopg2>=2.9",
]

[project.scripts]
tc-train = "transaction_classifier.training.cli:train"
tc-serve = "transaction_classifier.inference.cli:serve"
tc-hpo = "transaction_classifier.training.hpo.cli:hpo"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH", "G"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"src/transaction_classifier/core/features/*" = ["N806"]
"src/transaction_classifier/training/*" = ["N806"]
"src/transaction_classifier/inference/*" = ["N806"]
"src/transaction_classifier/core/data/splitter.py" = ["N806"]
"tests/unit/test_features.py" = ["N806"]
"tests/unit/test_hpo_objective.py" = ["N806"]
"tests/unit/test_hpo_cli.py" = ["N806"]
"tests/unit/test_predictor.py" = ["N806"]
"tests/unit/test_inference_app.py" = ["N806"]
"tests/unit/test_artifacts.py" = ["N806"]
"tests/unit/test_ops.py" = ["N806"]
"tests/unit/test_explain.py" = ["N806"]
"tests/unit/test_training_cli_extended.py" = ["N806"]
"tests/integration/test_pipeline.py" = ["N806"]
"tests/integration/test_api.py" = ["N806"]
"tests/integration/test_watcher.py" = ["N806"]
"scripts/*" = ["N806"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=transaction_classifier --cov-report=term-missing --cov-fail-under=80"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "joblib",
    "xgboost.*",
    "optuna.*",
    "optuna_integration.*",
    "shap",
    "scipy.*",
]
ignore_missing_imports = true
